requirements.txt
streamlit==1.30.0
streamlit-webrtc==0.47.0
streamlit-audiorecorder==0.0.6
transformers==4.36.0
torch>=1.13.0
requests>=2.28.0
python-dotenv>=1.0.0
pydub>=0.25.1
numpy>=1.24.0,<2.0.0


pandas>=2.0.0
openai>=1.3.0
huggingface-hub>=0.19.0
speechrecognition>=3.10.0
python-dateutil>=2.8.2
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.17.0
pathlib
json5
